+++ ts-sdk/cli/src/commands/fetch_tick_array.ts
@@ -1,49 +1,78 @@
 import { fetchMaybeFusionPool, fetchMaybeTickArray, getTickArrayAddress } from "@crypticdot/fusionamm-client";
-import BaseCommand, { addressArg } from "../base";
+import BaseCommand, { addressArg, addressFlag } from "../base";
 import { rpc } from "../rpc";
-import { _TICK_ARRAY_SIZE, getTickArrayStartTickIndex } from "@crypticdot/fusionamm-core";
-import { Args } from "@oclif/core";
+import { _TICK_ARRAY_SIZE, getTickArrayStartTickIndex, getTickIndexInArray } from "@crypticdot/fusionamm-core";
+import { Args, Flags } from "@oclif/core";
 
 export default class FetchTickArray extends BaseCommand {
   static override args = {
-    pool: addressArg({
-      description: "Fusion pool address",
-      required: true,
+    tickArray: addressArg({
+      description: "Tick array address",
     }),
-    tickIndex: Args.integer({
+  };
+  static override flags = {
+    pool: addressFlag({
+      description: "Fusion pool address",
+    }),
+    tickIndex: Flags.integer({
       description: "Tick index",
-      required: true,
     }),
   };
   static override description = "Fetch a tick array by tick index";
   static override examples = ["<%= config.bin %> <%= command.id %> 3qx1xPHwQopPXQPPjZDNZ4PnKpQvYeC3s8tPHcC5Ux1V 23000"];
 
   public async run() {
-    const { args } = await this.parse(FetchTickArray);
+    const { args, flags } = await this.parse(FetchTickArray);
 
-    console.log(`Fetching fusion pool at address ${args.pool}...`);
-    const fusionPool = await fetchMaybeFusionPool(rpc, args.pool);
-    if (!fusionPool.exists) {
-      throw new Error(`Fusion pool is not found at address ${args.pool}`);
-    }
-    console.log("Fusion pool:", fusionPool);
+    if (args.tickArray) {
+      console.log(`Fetching TickArray at address ${args.tickArray}...`);
+      const tickArray = await fetchMaybeTickArray(rpc, args.tickArray);
+      if (tickArray.exists) {
+        console.log(`Fetching fusion pool at address ${tickArray.data.fusionPool}...`);
+        const fusionPool = await fetchMaybeFusionPool(rpc, tickArray.data.fusionPool);
+        if (!fusionPool.exists) {
+          throw new Error(`Fusion pool is not found at address ${tickArray.data.fusionPool}`);
+        }
 
-    const startTickIndex = getTickArrayStartTickIndex(args.tickIndex, fusionPool.data.tickSpacing);
-    const tickArrayAddress = (await getTickArrayAddress(args.pool, startTickIndex))[0];
-
-    console.log(`Fetching TickArray at address ${tickArrayAddress}...`);
-    const tickArray = await fetchMaybeTickArray(rpc, tickArrayAddress);
-    if (tickArray.exists) {
-      console.log("TickArray", tickArray);
-      console.log("Ticks:");
-      for (let i = 0; i < _TICK_ARRAY_SIZE(); i++) {
-        console.log(
-          `Index ${i}, Tick ${tickArray.data.startTickIndex + i * fusionPool.data.tickSpacing}:`,
-          tickArray.data.ticks[i],
-        );
+        console.log("TickArray", tickArray);
+        for (let i = 0; i < _TICK_ARRAY_SIZE(); i++) {
+          console.log(
+            `Index=${i}, tickIndex=${tickArray.data.startTickIndex + i * fusionPool.data.tickSpacing}:`,
+            tickArray.data.ticks[i],
+          );
+        }
+      } else {
+        console.log("TickArray doesn't exist");
       }
     } else {
-      console.log("TickArray doesn't exist");
+      if (!flags.pool || !flags.tickIndex) {
+        throw new Error("Pool address and tickIndex must be provided if the tick array address is not set");
+      }
+
+      console.log(`Fetching fusion pool at address ${flags.pool}...`);
+      const fusionPool = await fetchMaybeFusionPool(rpc, flags.pool);
+      if (!fusionPool.exists) {
+        throw new Error(`Fusion pool is not found at address ${flags.pool}`);
+      }
+      console.log("Fusion pool:", fusionPool);
+
+      if (flags.tickIndex % fusionPool.data.tickSpacing != 0) {
+        throw new Error("The provided tickIndex must be initializable");
+      }
+      const startTickIndex = getTickArrayStartTickIndex(flags.tickIndex, fusionPool.data.tickSpacing);
+      const tickArrayAddress = (await getTickArrayAddress(flags.pool, startTickIndex))[0];
+
+      console.log(`Fetching TickArray at address ${tickArrayAddress}...`);
+      const tickArray = await fetchMaybeTickArray(rpc, tickArrayAddress);
+      if (tickArray.exists) {
+        console.log("TickArray", tickArray);
+
+        const i = getTickIndexInArray(flags.tickIndex, startTickIndex, fusionPool.data.tickSpacing);
+        console.log(`Tick ${flags.tickIndex}, index in array ${i}:`);
+        console.log(tickArray.data.ticks[i]);
+      } else {
+        console.log("TickArray doesn't exist");
+      }
     }
   }
 }