+++ ts-sdk/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @crypticdot/fusionamm-core
 
+## 1.0.54
+
 ## 1.0.53
 
 ## 1.0.52